[v3 Maintenance]: Consolidate and Document Core Changes in v3
